delete category

+++ resources/js/Pages/task/Task.vue
@@ -8,13 +8,18 @@ import CategoryList from '../../Components/CategoryList.vue';
 import CreateTask from '../../Components/CreateTask.vue';
 import ShowTask from '../../Components/ShowTask.vue';
 import SearchForm from '../../Components/SearchForm.vue';
+import { useRoute, useRouter } from 'vue-router';
 
 const tasks = ref([]);
 const taskDetail = ref({});
 const isLoading = ref(false);
 const openCreateForm = ref(false);
 const openTaskDetail = ref(false);
+const openSearchForm = ref(false);
 const errors = ref({});
+
+const route = useRoute();
+const router = useRouter();
 
 onMounted(() => {
   getTaskList();
@@ -23,21 +28,30 @@ onMounted(() => {
 const closeModal = () => {
   openCreateForm.value = false
   openTaskDetail.value = false
+  openSearchForm.value = false
 }
 
-const getTaskList = async form => {
+const getTaskList = async ()  => {
   try {
+    const id = route.params.id;
     isLoading.value = true
-    await axios.get('/api/task', { params: form })
-      .then(res => {
-        setTimeout(() => {
-          isLoading.value = false
-          tasks.value = res.data
-        }, 500);
-      });
+
+    let res = null;
+    if(id) {
+      res =  await axios.get(`/api/category/${route.params.id}/tasks`, {
+        params: route.query
+      })
+    }else {
+      res = await axios.get('/api/task', {
+        params: route.query
+      })
+    }
+    setTimeout(() => {
+      isLoading.value = false
+      tasks.value = res.data
+    }, 500);
   }catch(e) {
     isLoading.value = false
-    console.log(e)
   }
 }
 
@@ -75,6 +89,9 @@ const toggleStatus = async id => {
 }
 
 const deleteTask = async id => {
+  if(!confirm('本当に削除してもよろしいですか？')) {
+    return;
+  }
   try {
     await axios.delete(`/api/task/${id}`)
     isLoading.value = false
@@ -88,7 +105,7 @@ const deleteTask = async id => {
 
 <template>
   <auth-layout>
-    <!--カテゴリー-->
+    <!-- カテゴリ一覧 -->
     <CategoryList />
 
     <!-- タスク詳細 -->
@@ -96,15 +113,26 @@ const deleteTask = async id => {
 
     <!-- タスク作成フォーム -->
     <CreateTask :open="openCreateForm" :errors="errors" @closeModal="closeModal" @createTask="createTask" />
+    
+    <div class="lg:flex mt-8 w-full mx-auto overflow-auto">
+      <!--検索フォーム-->
+      <button @click="openSearchForm = !openSearchForm" :class="{'block': openSearchForm}" class="lg:hidden text-white bg-indigo-300 hover:bg-indigo-400 p-2 rounded">
+        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6 inline-block">
+          <path stroke-linecap="round" stroke-linejoin="round" d="M21 21l-5.197-5.197m0 0A7.5 7.5 0 105.196 5.196a7.5 7.5 0 0010.607 10.607z" />
+        </svg>
+        絞り込み
+      </button>
+      <SearchForm :open="openSearchForm" @search="getTaskList" />
 
-    <SearchForm @search="getTaskList" />
-
-    <TaskList :tasks="tasks" :isLoading="isLoading" @toggleStatus="toggleStatus" @deleteTask="deleteTask" />
+      <!--タスク一覧-->
+      <TaskList :tasks="tasks" :isLoading="isLoading" @showTask="showTask" @toggleStatus="toggleStatus" @deleteTask="deleteTask" />
+    </div>
 
     <button @click="openCreateForm = true" class="fixed bottom-10 right-5 text-white bg-indigo-400 hover:bg-indigo-500 p-4 rounded-full">
       <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-8 h-8">
         <path stroke-linecap="round" stroke-linejoin="round" d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L6.832 19.82a4.5 4.5 0 01-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 011.13-1.897L16.863 4.487zm0 0L19.5 7.125" />
       </svg>
     </button>
+
   </auth-layout>
 </template>
